streamlit_alynisit/requirements.txt
python-dotenv
altair
pandas
streamlit
pymongo
plotly
scipy